+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Add all service config to config file
   - Ensure tx_id is being passed in header
   - Change to use pytest to improve test output. Also improve code coverage stats
+  - Move logging bac in from sdx-common and remove the use of sdx-common
 
 ### 2.7.0 2017-10-16
   - Hardcode unchanging variables in settings.py to make configuration management simpler